refactor: extract component composition pattern from monolithic App.tsx
- Reduce App.tsx from 524 lines to 22 lines focused on providers
- Extract AppContainer for main app orchestration
- Create self-contained tab components with own event handling:
  - RecordingsTab: manages transcription history and recording events
  - ModelsTab: handles model management and downloads
  - SettingsTab: general settings with hotkey registration events
  - EnhancementsTab: AI enhancement features and API key management
  - AdvancedTab: advanced configuration options
  - AccountTab: license and account management
- Implement TabContainer with lazy loading and route management
- Each tab manages its own event listeners and state
- Maintain all existing functionality with better separation of concerns
