+++ src/components/tabs/AdvancedTab.tsx
@@ -1,0 +1,5 @@
+import { AdvancedSection } from "../sections/AdvancedSection";
+
+export function AdvancedTab() {
+  return <AdvancedSection />;
+}
