+++ src/components/tabs/index.ts
@@ -1,0 +1,8 @@
+// Export all tab components for clean imports
+export { RecordingsTab } from './RecordingsTab';
+export { ModelsTab } from './ModelsTab';
+export { SettingsTab } from './SettingsTab';
+export { EnhancementsTab } from './EnhancementsTab';
+export { AdvancedTab } from './AdvancedTab';
+export { AccountTab } from './AccountTab';
+export { TabContainer } from './TabContainer';
